[WIP] Move parsers/common_* to gorm

+++ packages/model/info_block.go
@@ -5,7 +5,7 @@ type InfoBlock struct {
 	StateID        int64  `gorm:"not null"`
 	WalletID       int64  `gorm:"not null"`
 	BlockID        int64  `gorm:"not null"`
-	Time           int32  `gorm:"not null"`
+	Time           int64  `gorm:"not null"`
 	Level          int8   `gorm:"not null"`
 	CurrentVersion string `gorm:"not null"`
 	Sent           int8   `gorm:"not null"`
